+++ src/momus_late.lua
@@ -1,6 +1,7 @@
 modutil.mod.Path.Context.Env("DoThrowEx", function (weaponName)
     modutil.mod.Path.Wrap("GetDerivedPropertyChangeValues", function (base, args)
         local derivedValues = base(args)
+		-- record projectile co-ords for momus duplication
         if args.ProjectileName == "ProjectileThrowCharged" and game.HeroHasTrait("StaffSelfHitAspect") and game.SessionMapState[_PLUGIN.guid.."DoThrowExRecord"] then
             table.insert(game.SessionMapState[_PLUGIN.guid.."DoThrowExRecord"], {
                 Angle = game.GetAngle({ Id = game.CurrentRun.Hero.ObjectId }),
@@ -8,6 +9,7 @@ modutil.mod.Path.Context.Env("DoThrowEx", function (weaponName)
                 Time = game._worldTime
             })
         end
+		-- trigger eos duplication on base projectile creation (they're created just after this call)
         if args.ProjectileName == "ProjectileThrowCharged" and game.HeroHasTrait("TorchSprintRecallAspect") and game.SessionMapState.CurrentExProjectile then
             local angle_record = game.GetAngle({ Id = game.CurrentRun.Hero.ObjectId })
             local projectileName = "ProjectileThrowCharged"
@@ -40,24 +42,26 @@ end)
 
 modutil.mod.Path.Context.Env("StartCastRepeatThread", function (triggerArgs, functionArgs)
 	modutil.mod.Path.Wrap("GetHeroTrait", function (base, traitName)
-		if traitName == "StaffSelfHitAspect" then
-			return base(traitName) or base(traitName.."_Secondary")
-		end
-		return base(traitName)
+		return GetHeroTraitWrap(base, traitName, "StaffSelfHitAspect")
 	end)
 end)
 
 modutil.mod.Path.Context.Env("WeaponCastFired", function (owner, weaponData, args, triggerArgs)
 	modutil.mod.Path.Wrap("HeroHasTrait", function (base, traitName)
-		if traitName == "StaffSelfHitAspect" then
-			return base(traitName) or base(traitName.."_Secondary")
-		end
-		return base(traitName)
+		return HeroHasTraitWrap(base, traitName, "StaffSelfHitAspect")
 	end)
 	modutil.mod.Path.Wrap("GetHeroTrait", function (base, traitName)
-		if traitName == "StaffSelfHitAspect" then
-			return base(traitName) or base(traitName.."_Secondary")
+		return GetHeroTraitWrap(base, traitName, "StaffSelfHitAspect")
+	end)
+end)
+
+modutil.mod.Path.Context.Env("StartSpecialRepeatThread", function (startX, startY, angle, functionArgs, triggerArgs)
+	modutil.mod.Path.Wrap("GetDerivedPropertyChangeValues", function (base, ...)
+		local derivedValues = base(...)
+		if game.HeroHasTrait("StaffRaiseDeadAspect_Secondary") or game.HeroHasTrait("StaffRaiseDeadAspect") then
+			derivedValues.PropertyChanges.DamageRadius = 435
+			derivedValues.PropertyChanges.Damage = 110
 		end
-		return base(traitName)
+		return derivedValues
 	end)
 end)